db_crud/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
